// src/enums/enums.ts
export enum ActivityLevel {
    SEDENTARY = 'SEDENTARY',
    LIGHT = 'LIGHT',
    MODERATE = 'MODERATE',
    ACTIVE = 'ACTIVE',
    VERY_ACTIVE = 'VERY_ACTIVE',
}

export enum Gender {
    MALE = 'MALE',
    FEMALE = 'FEMALE'
}

export enum Goal {
    GRADUAL_WEIGHT_LOSS = 'GRADUAL_WEIGHT_LOSS',
    MODERATE_WEIGHT_LOSS = 'MODERATE_WEIGHT_LOSS',
    EXTREME_WEIGHT_LOSS = 'EXTREME_WEIGHT_LOSS',
    MUSCLE_GAIN = 'MUSCLE_GAIN',
    MAINTAIN = 'MAINTAIN'
}

export enum Unit {
    METRIC = 'METRIC',
    IMPERIAL = 'IMPERIAL',
}

export enum CopyOrEditWorkout {
    COPY = 'COPY',
    EDIT = 'EDIT',
}

export enum Currency {
    AED = 'AED', // United Arab Emirates Dirham
    ARS = 'ARS', // Argentine Peso
    AUD = 'AUD', // Australian Dollar
    BRL = 'BRL', // Brazilian Real
    CAD = 'CAD', // Canadian Dollar
    CHF = 'CHF', // Swiss Franc
    CLP = 'CLP', // Chilean Peso
    CNY = 'CNY', // Chinese Yuan
    COP = 'COP', // Colombian Peso
    CZK = 'CZK', // Czech Koruna
    DKK = 'DKK', // Danish Krone
    EUR = 'EUR', // Euro
    GBP = 'GBP', // British Pound Sterling
    HKD = 'HKD', // Hong Kong Dollar
    HUF = 'HUF', // Hungarian Forint
    IDR = 'IDR', // Indonesian Rupiah
    ILS = 'ILS', // Israeli New Shekel
    INR = 'INR', // Indian Rupee
    JPY = 'JPY', // Japanese Yen
    KRW = 'KRW', // South Korean Won
    MXN = 'MXN', // Mexican Peso
    MKD = 'MKD', // Macedonian Denar
    MYR = 'MYR', // Malaysian Ringgit
    NOK = 'NOK', // Norwegian Krone
    NZD = 'NZD', // New Zealand Dollar
    PHP = 'PHP', // Philippine Peso
    PLN = 'PLN', // Polish Zloty
    RUB = 'RUB', // Russian Ruble
    SEK = 'SEK', // Swedish Krona
    SGD = 'SGD', // Singapore Dollar
    THB = 'THB', // Thai Baht
    TWD = 'TWD', // New Taiwan Dollar
    TRY = 'TRY', // Turkish Lira
    USD = 'USD', // United States Dollar
    ZAR = 'ZAR'  // South African Rand
}

export enum VenueType {
    GYM = 'GYM',
    SHOP = 'SHOP',
    RESTAURANT = 'RESTAURANT'
}

export enum FoodCategory {
    BREAKFAST = 'BREAKFAST',
    MAIN_DISHES = 'MAIN_DISHES',
    SNACK = 'SNACK',
    DRINKS = 'DRINKS'
}

export enum MealType {
    BREAKFAST = 'BREAKFAST',
    LUNCH = 'LUNCH',
    DINNER = 'DINNER',
    SNACK = 'SNACK',
}

export enum ExerciseCategory {
    STRENGTH = 'STRENGTH',
    STRETCHING = 'STRETCHING',
    PLYOMETRICS = 'PLYOMETRICS',
    STRONGMAN = 'STRONGMAN',
    POWERLIFTING = 'POWERLIFTING',
    CARDIO = 'CARDIO',
    OLYMPIC_WEIGHTLIFTING = 'OLYMPIC_WEIGHTLIFTING',
}

export enum Force {
    PULL = 'PULL',
    PUSH = 'PUSH',
    STATIC = 'STATIC',
    ELASTIC = 'ELASTIC',
    ROTATIONAL = 'ROTATIONAL',
    EXPLOSIVE = 'EXPLOSIVE',
}

export enum Mechanic {
    COMPOUND = 'COMPOUND',
    ISOLATION = 'ISOLATION',
}

export enum ExerciseLevel {
    BEGINNER = 'BEGINNER',
    INTERMEDIATE = 'INTERMEDIATE',
    EXPERT = 'EXPERT',
}